requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.2
bpyutils==0.5.8
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
DateTime==5.4
decorator==5.1.1
h11==0.14.0
idna==3.6
numpy==2.1.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
pipupgrade==1.12.0
py==1.11.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
retry==0.9.2
selenium==4.16.0
setuptools==69.0.3
six==1.16.0
slack-bolt==1.18.1
slack_sdk==3.26.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.24.0
trio-websocket==0.11.1
tzdata==2024.1
urllib3==2.1.0
webdriver-manager==4.0.1
wsproto==1.2.0
zope.interface==6.1
